Refactor: Extract new mod `logging`
