+++ crates/bin/src/lib.rs
@@ -2,4 +2,5 @@ pub mod args;
 pub mod bin_util;
 pub mod entry;
 pub mod install_path;
+pub mod logging;
 pub mod ui;
