+++ crates/bin/src/main.rs
@@ -6,7 +6,8 @@ use log::debug;
 use cargo_binstall::{
     args,
     bin_util::{run_tokio_main, MainExit},
-    entry, ui,
+    entry,
+    logging::logging,
 };
 
 #[cfg(feature = "mimalloc")]
@@ -26,7 +27,7 @@ fn main() -> MainExit {
         println!("{}", env!("CARGO_PKG_VERSION"));
         MainExit::Success(None)
     } else {
-        let _guard = ui::logging(&args);
+        let _guard = logging(&args);
 
         let start = Instant::now();
